feat(catalog): faster category product count (#25)
Previously, in the admin panel, computing the CategoryTree UI (the left-hand
element of the page that loads when you try to view the categories
of your store) was wildly slow for the root scope when there were
many `is_anchor` categories. This was due to several bugs:

1: Use of a missing index: `catalog_category_product_index` that is now defunct
and always empty. It was replaced by `catalog_category_product_index_store*`.
However, this table did not exist for the `store0`. This diff both adds
that missing table and changes the collection to use the new tables.

2. A bad computation of the `rootCategoryIds` for the root store, causing
a bad join condition against the temporary category recursion table
resulting in the resulting index table (catalog_category_product_index_store0)
always being empty.

3. An exceedingly slow private method (getProductsCountFromCategoryTable)
which should not have been merged in the first place given how slow it is.
I have removed this method. Removing this does cause a small UX bug
when the index is empty: the category page will show 0 products before
an index has run. It's either this or perform the computation on
every page refresh. I see this an acceptable break, given how slow
it is to attempt to compute these values manually.

Co-authored-by: Vincent ENJALBERT <chef@web-cooking.net>