Create category-product index table for store 0
This fixes a regression introduced by pull request
https://github.com/mage-os/mageos-magento2/pull/25

Under certain conditions saving or loading a product caused a
table-not-exists exception.
This happened for example for the integration test
\Magento\AdvancedPricingImportExport\Model\Import\AdvancedPricingTest::testImportDelete

With this change, the index table is created by the schema patch, just
like all other store index tables, and the error no longer occurs.

Closes #44