Use mage-os as vendor-name for composer info (#48)
The ComposerInformation class is used to extract information about the
current installation from the project root composer.json and
composer.lock files, for example the Magento version.
To find the relevant information, it looks at the package names, using a
vendor name magento.
Without this change, the installed version is reported as "UNKNOWN" in
the Magento admin when the mage-os/project-community-edition package is
used. This change fixes the regex so it correctly identifies the
installed version.

In future, it may be worth generalizing this so it is easier to switch
the project vendor for other distributions based on Mage-OS.

Fixes issue #42.