Remove AdminAnalytics Adobe tracking URL
This change ensures that in case the tracking-feature is enabled
in the admin, Mage-OS will not be tracked as Magento Open Source
or Enterprise Edition installations incorrectly.

In order to use admin analytics, a custom tracking_url script needs to
be configured on the tracking block.

Fixes https://github.com/mage-os/mageos-magento2/issues/43